refactor homepage Fixes #74
