+++ package.json
@@ -13,7 +13,7 @@
     "firebase": "^10.13.1",
     "firebase-admin": "^12.4.0",
     "formik": "^2.4.6",
-    "next": "14.2.4",
+    "next": "^14.2.12",
     "react": "^18",
     "react-dom": "^18",
     "yup": "^1.4.0"
@@ -24,6 +24,7 @@
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "^0.2.4",
+    "eslint-plugin-react": "^7.36.1",
     "typescript": "^5"
   }
 }
