+++ src/app/projects/page.tsx
@@ -1,11 +1,3 @@
-import { Layout, Maintenance } from '@/libs/ui';
+import { ProjectPage } from '@/libs/ui';
 
-const Projects = () => {
-  return (
-    <Layout>
-      <Maintenance pageTitle='Projects' />
-    </Layout>
-  );
-};
-
-export default Projects;
+export default ProjectPage;
