--- src/libs/ui/molecule/herobox/index.ts
@@ -1,1 +1,0 @@
-export {default as HeroBox} from './herobox';
