+++ src/libs/ui/molecule/index.ts
@@ -1,5 +1,4 @@
 export { NavBar } from './nav';
-export { HeroBox } from './herobox';
 export { Socials } from './socials';
 
 // to change later
