+++ src/libs/ui/organism/home/head/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeHeader } from './head';
