+++ src/libs/ui/organism/index.ts
@@ -1,1 +1,7 @@
+//home
+export {default as HomeHeader} from './home/head/head';
 
+//blog
+//projects
+//services
+//contact
