+++ src/libs/ui/pages/admin/page.tsx
@@ -1,0 +1,7 @@
+import { Maintenance } from '@/libs/ui';
+
+const AdminPage = () => {
+  return <Maintenance pageTitle='Admin' />;
+};
+
+export default AdminPage;
