+++ src/libs/ui/pages/blog/page.tsx
@@ -1,0 +1,11 @@
+import { Layout, Maintenance } from '@/libs/ui';
+
+const BlogPage = () => {
+  return (
+    <Layout>
+      <Maintenance pageTitle='Blog' />
+    </Layout>
+  );
+};
+
+export default BlogPage;
