+++ src/libs/ui/pages/contact/page.tsx
@@ -1,0 +1,11 @@
+import { Layout, Maintenance } from '@/libs/ui';
+
+const ContactPage = () => {
+  return (
+    <Layout>
+      <Maintenance pageTitle='Contact' />
+    </Layout>
+  );
+};
+
+export default ContactPage;
