+++ src/libs/ui/pages/home/page.tsx
@@ -1,0 +1,12 @@
+import { Layout, Maintenance } from '@/libs/ui';
+import { HomeHeader } from '../../organism';
+
+const HomePage = () => {
+  return (
+    <Layout>
+      <HomeHeader />
+    </Layout>
+  );
+};
+
+export default HomePage;
