+++ src/libs/ui/pages/index.ts
@@ -1,0 +1,6 @@
+export { default as AdminPage } from './admin/page';
+export { default as BlogPage } from './blog/page';
+export { default as ContactPage } from './contact/page';
+export { default as HomePage } from './home/page';
+export { default as ProjectPage } from './projects/page';
+export { default as ServicePage } from './services/page';
