+++ src/libs/ui/pages/projects/page.tsx
@@ -1,0 +1,11 @@
+import { Layout, Maintenance } from '@/libs/ui';
+
+const Projects = () => {
+  return (
+    <Layout>
+      <Maintenance pageTitle='Projects' />
+    </Layout>
+  );
+};
+
+export default Projects;
