+++ src/libs/ui/pages/services/page.tsx
@@ -1,0 +1,11 @@
+import { Layout, Maintenance } from '@/libs/ui';
+
+const ServicesPage = () => {
+  return (
+    <Layout>
+      <Maintenance pageTitle='Services' />
+    </Layout>
+  );
+};
+
+export default ServicesPage;
